requirements.txt
django==1.11
Pillow==5.1.0
django-pagedown==1.0.4
django-markdown-deux==1.0.5
django-crispy-forms==1.7.2
djangorestframework==3.8.2
markdown==2.6.11
django-filter==1.1.0
django-tracking
